feat: :sparkles: 2. Tipagem estaticas e criação de metodos
