+++ app/models/negociacao.ts
@@ -1,27 +1,27 @@
 export class Negociacao {
-  private _data;
-  private _quantidade;
-  private _valor;
+  private _data: Date;
+  private _quantidade: number;
+  private _valor: number;
 
-  constructor(data, quantidade, valor) {
+  constructor(data: Date, quantidade: number, valor: number) {
     this._data = data;
     this._quantidade = quantidade;
     this._valor = valor;
   }
 
-  get data() {
+  get data(): Date {
     return this._data;
   }
 
-  get quantidade() {
+  get quantidade(): number {
     return this._quantidade;
   }
 
-  get valor() {
+  get valor(): number {
     return this._valor;
   }
 
-  get volume() {
+  get volume(): number {
     return this._quantidade * this._valor;
   }
 }
