+++ tsconfig.json
@@ -2,7 +2,8 @@
   "compilerOptions": {
     "outDir": "dist/js",
     "target": "ES6",
-    "noEmitOnError": true
+    "noEmitOnError": true,
+    "noImplicitAny": true
   },
   "include": ["app/**/*"]
 }
